Digital Art Gallery with No-Code E-commerce
This project started with the goal of simplifying the creation of an e-commerce website for art galleries, using a no-code approach. The key advantage of this strategy is the speed and ease of implementation, allowing anyone to create a professional platform without writing a single line of code. I've outlined the main ideas and necessary features for the project's success: from organizing the artwork catalog to integrating secure payment tools and personalizing profiles for artists and clients. ArtConnect reflects my vision of merging art and technology, making art more accessible in the digital space.